Move auth pages to a subfolder
